ES-1847: topic-config
