+++ libs/corda-sdk/build.gradle
@@ -13,6 +13,8 @@ ext {
 
 dependencies {
     compileOnly "net.corda.cli.host:api:$pluginHostVersion"
+    compileOnly libs.jackson.dataformat.yaml
+
     implementation project(':components:membership:membership-rest')
     implementation libs.jackson.dataformat.yaml
     implementation libs.jackson.databind
