+++ tools/plugins/topic-config/build.gradle
@@ -24,6 +24,7 @@ dependencies {
     compileOnly platform("net.corda:corda-api:$cordaApiVersion")
     implementation "net.corda:corda-topic-schema:$cordaApiVersion"
     implementation libs.kafka.client
+    implementation project(':libs:corda-sdk')
     constraints {
         implementation(libs.slf4j.v2.api)
     }
